requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0 